Prevent garbage collection from occurring in module functions
It seems garbage collection deleted references created in the dynamic module before they were passed to
the Elisp code, which made Emacs try to access addresses that were out of bounds (because they were
gc'ed) when running Elisp code and crashed it.

Also make a global reference to all global interns used in the module and initialized at the beginning.